Test: add coverage for character filtering and sorting
