+++ src/__tests__/App.test.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { App } from "../components/App";
+
+it("renders without crashing", () => {
+    const div = document.createElement("div");
+    const root = createRoot(div!);
+    root.render(<App />);
+    root.unmount();
+});
